First pass at eliminating passthru/auth errors.
Skip the authentication before action and rely instead on Apache rules
restricting pages to Shibboleth login.

This approach still fails to force a check of the Shibboleth session to create
a Rails session and often dumps users to the homepage, but at least that's
within the app and worked around with a refresh.