Covid: Restore Lending and laser hours on homepage.
These services are returning to a normal access model for Fall 2021,
albeit beginning with reduced hours due to student staffing levels, so
bring these back to a place of prominence on the homepage.

Keep Wood Shop hours hidden for now, as that is not likely to see
regular staffing in the near future.  (It remains linked from main
navigation.)

This partially reverts commit 3556055b870c6fa87834c71accdadc563e1b8b28.